requirements.txt
torch~=1.8.1
torchvision~=0.9.1
Pillow~=8.2.0
nltk~=3.6.2
tqdm~=4.61.0
pytorch_lightning~=1.3.5
h5py~=3.2.1
numpy~=1.20.3
PyYAML~=5.4.1